part4 - ex4.21 - user auth on delete + restrictions

+++ part4/bloglist/controllers/blogs.js
@@ -35,6 +35,17 @@ blogsRouter.post("/", async (request, response) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const blog = await Blog.findById(request.params.id);
+
+    if(decodedToken.id.toString() !== blog.user.toString()){
+        return response.status(401).json({ error: 'user unauthorized to delete this blog' })
+    }
+
     await Blog.findByIdAndRemove(request.params.id);
     response.status(204).end();
 });
